feat: hot reload (#3392)
JVM languages can now used how reload for instant feedback

+++ backend/runner/runner.go
@@ -34,6 +34,7 @@ import (
 	"github.com/TBD54566975/ftl/backend/runner/observability"
 	"github.com/TBD54566975/ftl/common/plugin"
 	"github.com/TBD54566975/ftl/internal/download"
+	"github.com/TBD54566975/ftl/internal/exec"
 	"github.com/TBD54566975/ftl/internal/identity"
 	"github.com/TBD54566975/ftl/internal/log"
 	"github.com/TBD54566975/ftl/internal/model"
@@ -59,6 +60,7 @@ type Config struct {
 	DebugPort             int                      `help:"The port to use for debugging." env:"FTL_DEBUG_PORT"`
 	Registry              artefacts.RegistryConfig `embed:"" prefix:"oci-"`
 	ObservabilityConfig   ftlobservability.Config  `embed:"" prefix:"o11y-"`
+	DevEndpoint           optional.Option[url.URL] `help:"An existing endpoint to connect to in development mode" env:"FTL_DEV_ENDPOINT"`
 }
 
 func Start(ctx context.Context, config Config) error {
@@ -125,6 +127,7 @@ func Start(ctx context.Context, config Config) error {
 		labels:             labels,
 		deploymentLogQueue: make(chan log.Entry, 10000),
 		cancelFunc:         doneFunc,
+		devEndpoint:        config.DevEndpoint,
 	}
 	err = svc.deploy(ctx)
 	if err != nil {
@@ -242,10 +245,12 @@ func manageDeploymentDirectory(logger *log.Logger, config Config) error {
 var _ ftlv1connect.VerbServiceHandler = (*Service)(nil)
 
 type deployment struct {
-	key    model.DeploymentKey
-	plugin *plugin.Plugin[ftlv1connect.VerbServiceClient]
+	key model.DeploymentKey
 	// Cancelled when plugin terminates
-	ctx context.Context
+	ctx      context.Context
+	cmd      optional.Option[exec.Cmd]
+	endpoint *url.URL // The endpoint the plugin is listening on.
+	client   ftlv1connect.VerbServiceClient
 }
 
 type Service struct {
@@ -262,6 +267,7 @@ type Service struct {
 	labels              *structpb.Struct
 	deploymentLogQueue  chan log.Entry
 	cancelFunc          func()
+	devEndpoint         optional.Option[url.URL]
 }
 
 func (s *Service) Call(ctx context.Context, req *connect.Request[ftlv1.CallRequest]) (*connect.Response[ftlv1.CallResponse], error) {
@@ -269,7 +275,7 @@ func (s *Service) Call(ctx context.Context, req *connect.Request[ftlv1.CallReque
 	if !ok {
 		return nil, connect.NewError(connect.CodeUnavailable, errors.New("no deployment"))
 	}
-	response, err := deployment.plugin.Client.Call(ctx, req)
+	response, err := deployment.client.Call(ctx, req)
 	if err != nil {
 		deploymentLogger := s.getDeploymentLogger(ctx, deployment.key)
 		deploymentLogger.Errorf(err, "Call to deployments %s failed to perform gRPC call", deployment.key)
@@ -339,41 +345,53 @@ func (s *Service) deploy(ctx context.Context) error {
 			return fmt.Errorf("failed to create deployment directory: %w", err)
 		}
 	}
-	storage, err := artefacts.NewOCIRegistryStorage(s.config.Registry)
-	if err != nil {
-		return fmt.Errorf("failed to create OCI registry storage: %w", err)
-	}
-	err = download.ArtefactsFromOCI(ctx, s.controllerClient, key, deploymentDir, storage)
-	if err != nil {
-		observability.Deployment.Failure(ctx, optional.Some(key.String()))
-		return fmt.Errorf("failed to download artefacts: %w", err)
+	var dep *deployment
+	if ep, ok := s.devEndpoint.Get(); ok {
+		client := rpc.Dial(ftlv1connect.NewVerbServiceClient, ep.String(), log.Error)
+		dep = &deployment{
+			ctx:      ctx,
+			key:      key,
+			cmd:      optional.None[exec.Cmd](),
+			endpoint: &ep,
+			client:   client,
+		}
+	} else {
+		storage, err := artefacts.NewOCIRegistryStorage(s.config.Registry)
+		if err != nil {
+			return fmt.Errorf("failed to create OCI registry storage: %w", err)
+		}
+		err = download.ArtefactsFromOCI(ctx, s.controllerClient, key, deploymentDir, storage)
+		if err != nil {
+			observability.Deployment.Failure(ctx, optional.Some(key.String()))
+			return fmt.Errorf("failed to download artefacts: %w", err)
+		}
+
+		envVars := []string{"FTL_ENDPOINT=" + s.config.ControllerEndpoint.String(),
+			"FTL_CONFIG=" + strings.Join(s.config.Config, ","),
+			"FTL_OBSERVABILITY_ENDPOINT=" + s.config.ControllerEndpoint.String()}
+		if s.config.DebugPort > 0 {
+			envVars = append(envVars, fmt.Sprintf("FTL_DEBUG_PORT=%d", s.config.DebugPort))
+		}
+
+		verbCtx := log.ContextWithLogger(ctx, deploymentLogger.Attrs(map[string]string{"module": module.Name}))
+		deployment, cmdCtx, err := plugin.Spawn(
+			unstoppable.Context(verbCtx),
+			log.FromContext(ctx).GetLevel(),
+			gdResp.Msg.Schema.Name,
+			deploymentDir,
+			"./launch",
+			ftlv1connect.NewVerbServiceClient,
+			plugin.WithEnvars(
+				envVars...,
+			),
+		)
+		if err != nil {
+			observability.Deployment.Failure(ctx, optional.Some(key.String()))
+			return fmt.Errorf("failed to spawn plugin: %w", err)
+		}
+		dep = s.makeDeployment(cmdCtx, key, deployment)
 	}
 
-	envVars := []string{"FTL_ENDPOINT=" + s.config.ControllerEndpoint.String(),
-		"FTL_CONFIG=" + strings.Join(s.config.Config, ","),
-		"FTL_OBSERVABILITY_ENDPOINT=" + s.config.ControllerEndpoint.String()}
-	if s.config.DebugPort > 0 {
-		envVars = append(envVars, fmt.Sprintf("FTL_DEBUG_PORT=%d", s.config.DebugPort))
-	}
-
-	verbCtx := log.ContextWithLogger(ctx, deploymentLogger.Attrs(map[string]string{"module": module.Name}))
-	deployment, cmdCtx, err := plugin.Spawn(
-		unstoppable.Context(verbCtx),
-		log.FromContext(ctx).GetLevel(),
-		gdResp.Msg.Schema.Name,
-		deploymentDir,
-		"./launch",
-		ftlv1connect.NewVerbServiceClient,
-		plugin.WithEnvars(
-			envVars...,
-		),
-	)
-	if err != nil {
-		observability.Deployment.Failure(ctx, optional.Some(key.String()))
-		return fmt.Errorf("failed to spawn plugin: %w", err)
-	}
-
-	dep := s.makeDeployment(cmdCtx, key, deployment)
 	s.readyTime.Store(time.Now().Add(time.Second * 2)) // Istio is a bit flakey, add a small delay for readiness
 	s.deployment.Store(optional.Some(dep))
 	logger.Debugf("Deployed %s", key)
@@ -395,19 +413,21 @@ func (s *Service) Close() error {
 	if !ok {
 		return connect.NewError(connect.CodeNotFound, errors.New("no deployment"))
 	}
-	// Soft kill.
-	err := depl.plugin.Cmd.Kill(syscall.SIGTERM)
-	if err != nil {
-		return fmt.Errorf("failed to kill plugin: %w", err)
-	}
-	// Hard kill after 10 seconds.
-	select {
-	case <-depl.ctx.Done():
-	case <-time.After(10 * time.Second):
-		err := depl.plugin.Cmd.Kill(syscall.SIGKILL)
+	if cmd, ok := depl.cmd.Get(); ok {
+		// Soft kill.
+		err := cmd.Kill(syscall.SIGTERM)
 		if err != nil {
-			// Should we os.Exit(1) here?
 			return fmt.Errorf("failed to kill plugin: %w", err)
+		}
+		// Hard kill after 10 seconds.
+		select {
+		case <-depl.ctx.Done():
+		case <-time.After(10 * time.Second):
+			err := cmd.Kill(syscall.SIGKILL)
+			if err != nil {
+				// Should we os.Exit(1) here?
+				return fmt.Errorf("failed to kill plugin: %w", err)
+			}
 		}
 	}
 	s.deployment.Store(optional.None[*deployment]())
@@ -421,16 +441,18 @@ func (s *Service) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, "no deployment", http.StatusNotFound)
 		return
 	}
-	proxy := httputil.NewSingleHostReverseProxy(deployment.plugin.Endpoint)
+	proxy := httputil.NewSingleHostReverseProxy(deployment.endpoint)
 	proxy.ServeHTTP(w, r)
 
 }
 
 func (s *Service) makeDeployment(ctx context.Context, key model.DeploymentKey, plugin *plugin.Plugin[ftlv1connect.VerbServiceClient]) *deployment {
 	return &deployment{
-		ctx:    ctx,
-		key:    key,
-		plugin: plugin,
+		ctx:      ctx,
+		key:      key,
+		cmd:      optional.Ptr(plugin.Cmd),
+		endpoint: plugin.Endpoint,
+		client:   plugin.Client,
 	}
 }
